Split widget: panel placeholder
